+++ src/auth/auth.controller.ts
@@ -1,14 +1,4 @@
-import {
-  Body,
-  Controller,
-  Get,
-  Param,
-  Post,
-  Query,
-  Res,
-  UseGuards,
-} from '@nestjs/common';
-import { AuthGuard } from '@nestjs/passport';
+import { Body, Controller, Param, Post, UseGuards } from '@nestjs/common';
 import {
   ApiBadRequestResponse,
   ApiBearerAuth,
@@ -16,16 +6,13 @@ import {
   ApiNotFoundResponse,
   ApiOkResponse,
   ApiOperation,
-  ApiResponse,
   ApiTags,
   ApiUnauthorizedResponse,
 } from '@nestjs/swagger';
-import { Response } from 'express';
 import { ErrorOutput } from '../common/dtos/output.dto';
 import { User } from '../users/entities/user.entity';
 import { AuthUser } from './auth-user.decorator';
-import { AuthService, OauthService } from './auth.service';
-import { googleUserInfo } from './dtos/google.dto';
+import { AuthService } from './auth.service';
 import {
   LoginBodyDto,
   LoginOutput,
@@ -126,82 +113,5 @@ export class AuthController {
     @Param('email') email: string,
   ): Promise<sendPasswordResetEmailOutput> {
     return this.authService.sendPasswordResetEmail(email);
-  }
-}
-
-@Controller('oauth')
-@ApiTags('OAuth')
-export class OauthController {
-  constructor(private readonly oauthService: OauthService) {}
-
-  @ApiOperation({
-    summary: '카카오 계정 로그인 요청',
-    description: '카카오 계정 로그인 요청 메서드',
-  })
-  @ApiResponse({
-    description: '카카오 계정 로그인 창이 켜지는 Redirect URL을 반환한다.',
-    status: 302,
-  })
-  @Get('kakao-auth')
-  async kakaoAuthorize(@Res() res: Response): Promise<void> {
-    const { url } = await this.oauthService.kakaoAuthorize();
-    return res.redirect(url);
-  }
-
-  @ApiOperation({
-    summary: '카카오 로그인',
-    description:
-      '카카오 로그인 메서드. (회원가입이 안되어 있으면 회원가입 처리 후 로그인 처리)',
-  })
-  @ApiOkResponse({
-    description: '로그인 성공 여부와 함께 access, refresh token을 반환한다.',
-    type: LoginOutput,
-  })
-  @ApiBadRequestResponse({
-    description:
-      '카카오 로그인 요청 시 발생하는 에러를 알려준다.(ex : email 제공에 동의하지 않은 경우)',
-    type: ErrorOutput,
-  })
-  @ApiUnauthorizedResponse({
-    description: '카카오 로그인 실패 여부를 알려준다.',
-    type: ErrorOutput,
-  })
-  @Get('kakao-login')
-  async kakaoOauth(@Query('code') code: string): Promise<LoginOutput> {
-    console.log(code);
-    return this.oauthService.kakaoOauth({ code });
-  }
-
-  @ApiOperation({
-    summary: '구글 계정 로그인 요청',
-    description: '구글 계정 로그인 요청 메서드',
-  })
-  @ApiResponse({
-    description: '구글 계정 로그인 창이 켜지는 Redirect URL을 반환한다.',
-    status: 302,
-  })
-  @Get('google-auth')
-  @UseGuards(AuthGuard('google'))
-  googleOauth(): void {}
-
-  @ApiOperation({
-    summary: '구글 로그인',
-    description:
-      '구글 로그인 메서드. (회원가입이 안되어 있으면 회원가입 처리 후 로그인 처리)',
-  })
-  @ApiOkResponse({
-    description: '로그인 성공 여부와 함께 access, refresh token을 반환한다.',
-    type: LoginOutput,
-  })
-  @ApiBadRequestResponse({
-    description: 'code가 잘못된 경우',
-    type: ErrorOutput,
-  })
-  @Get('google-login')
-  @UseGuards(AuthGuard('google'))
-  async googleAuthRedirect(
-    @AuthUser() user: googleUserInfo,
-  ): Promise<LoginOutput> {
-    return this.oauthService.googleOauth(user);
   }
 }
